COMP: speed up superbuild compilation - only call git svn info if needed
When a git clone of slicer is not connected to svn, then
git svn info can take a long time to run (needs several network
timeouts or something).  This new version checks to see if the string
svn-remote shows up in the git config and only if it is there will
the git svn info command be invoked.