BUG: Default *_WC_ROOT to 'NA' when no git origin is set in Extension
If the repository of an extension didn't have a remote.origin.url set,
the following unhelpful error message was being displayed:

CMake Error at
Slicer/Extensions/CMake/SlicerFunctionGenerateExtensionDescription.cmake:31
(message):
  CMake variable EXTENSION_WC_URL is empty !
Call Stack (most recent call first):
  Slicer/CMake/SlicerExtensionCPack.cmake:54
  (slicerFunctionGenerateExtensionDescription)
  CMakeLists.txt:64 (include)

This patch keeps *_WC_ROOT, and *_WC_URL set to "NA" if the origin is
not set in the repository. Also a CMake warning message is displayed
to indicate that the repository may not be configured correctly.